+++ src/core/syswarden-cli/cmd/tui_flush_linux.go
@@ -1,0 +1,12 @@
+//go:build linux
+
+package cmd
+
+import (
+	"os"
+	"golang.org/x/sys/unix"
+)
+
+func flushStdin() {
+	_ = unix.IoctlSetInt(int(os.Stdin.Fd()), unix.TCFLSH, unix.TCIFLUSH)
+}