+++ src/core/syswarden-cli/cmd/tui_flush_others.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package cmd
+
+func flushStdin() {
+	// Not required or not supported
+}